% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ATP_W144_excerpt}
\alias{ATP_W144_excerpt}
\title{Pew ATP Wave 144 Excerpt}
\format{
A data frame with 626 rows and several variables. See codebook for details.
}
\source{
"American Trends Panel: Wave 144." Pew Research Center, Washington, D.C. (March 18-24, 2024).
}
\usage{
ATP_W144_excerpt
}
\description{
A teaching excerpt from Pew Research Center's American Trends Panel Wave 144 (March 2024).
}
\details{
Pew Research Center bears no responsibility for the analyses or interpretations of the data presented here. The opinions expressed herein, including any implications for policy, are those of the author and not of Pew Research Center.
}
\keyword{datasets}
