% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_project.R
\name{final_project}
\alias{final_project}
\title{Create a final project folder structure for a given class}
\usage{
final_project(class_no)
}
\arguments{
\item{class_no}{Character. The class number, e.g., "mc451" or "mc501".}
}
\value{
Invisibly returns the path to the created project folder.
}
\description{
Create a final project folder structure for a given class
}
