% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simpsons_episodes}
\alias{simpsons_episodes}
\title{simpsons_episodes}
\format{
A data frame with columns:
\describe{
\item{id}{Unique identifier for the episode (integer)}
\item{title}{Episode title (character)}
\item{season}{Season number (integer)}
\item{number_in_season}{Episode number within the season (integer)}
\item{number_in_series}{Episode number within the series (integer)}
\item{original_air_date}{Original air date (Date)}
\item{original_air_year}{Original air year (integer)}
\item{production_code}{Production code (character)}
\item{imdb_rating}{IMDB rating (numeric)}
\item{imdb_votes}{Number of IMDB votes (integer)}
\item{us_viewers_in_millions}{US viewers in millions (numeric)}
\item{views}{Total views (numeric)}
\item{image_url}{URL of the episode image (character)}
\item{video_url}{URL of the episode video (character)}
}
}
\source{
Generated for educational purposes.
}
\usage{
simpsons_episodes
}
\description{
A dataset containing information about episodes of The Simpsons.
}
\keyword{datasets}
