\name{w144_excerpt}
\alias{w144_excerpt}
\docType{data}
\title{Excerpt of Pew Research ATP W144 Data}
\description{
  A cleaned excerpt of Pew Research Center's ATP W144 survey data, used for MC451/MC501 labs.
}
\format{A data frame with 1000 rows and [number] columns.}
\source{Pew Research Center, "ATP W144: Politics and News on Social Media Platforms." Washington, D.C. (March 2024).}